refactor(wal): widen WAL record magic to 8-byte BEACHWAL and add named layout offset constants

+++ internal/wal/reader.go
@@ -84,7 +84,7 @@ func (r *Reader) Next() (payload []byte, err error) {
 		return nil, err
 	}
 
-	r.pos += recordHeaderSize + int64(payloadLen)
+	r.pos += int64(recordHeaderSize) + int64(payloadLen)
 
 	return payload, nil
 }
